python/simple_ingestion_and_search/requirements.txt
shilp-sdk>=0.11.1
